worker/requirements.txt
fastapi
uvicorn
langchain-community
langchain_openai
langchain
openai
langchain_pinecone
pinecone
python-dotenv
PyMuPDF
pandas
lxml
python-multipart